third_party/WebKit/LayoutTests/editing/pasteboard/set_data_typeof_return.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<script src="../../resources/js-test.js"></script>
<script>
function setDataTest() {
    shouldBeUndefined("event.clipboardData.setData('text/plain', 'copy test')");
}
</script>
</head>
<body oncopy="setDataTest()">
<script>
description("Simple test to validate the return type of setData(). To run the test manually, simply select any text and initiate a copy operation.");

document.execCommand("copy");
</script>
</body>
</html>
